+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - `AdaptiveGridRouting.AddRoutingFilter(RoutingFilter f)`
 - `AdaptiveGraphRouting.RoutingConfiguration.SupportedAngles` property.
 - Default values for `AdaptiveGraphRouting.RoutingConfiguration` constructor.
+- `Line.BestFit(IList<Vector3> points)`
+- `Vector3Extensions.BestFitLine(this IList<Vector3> points)`
+- `Polygon.FromAlignedBoundingBox2d(IEnumerable<Vector3> points, Vector3 axis, double minSideSize = 0.1)`
 
 ### Changed